CarApp/python_app/pyproject.toml
[tool.poetry]
name = "python_app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.54.0"
grpcio-tools = "^1.54.0"
protobuf = "^4.23.4"
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
